rest of the SQL queries fleshed out
case 1 #2 uses TA ids which i didnt think we were using and aren't really necessary. Im ok with using them, just wanted to bring this up. Also, we need to add a course_crn field to workshopApplications because there is currently no way to learn a cowrkshop and its crn to what course it is for